move useRestorePassword()
