+++ projects/graphql-client/src/index.ts
@@ -42,3 +42,4 @@ export * from "./remote/mutations/useLogin"
 export * from "./remote/mutations/useUpdateAuth"
 export * from "./remote/mutations/useLogout"
 export * from "./remote/mutations/useIssueAuthToken"
+export * from "./remote/mutations/useRestorePassword"
